requirements.txt
websockets==10.1
pycryptodome==3.12.0
